Move HOCs to sub folder
