+++ assets/src/components/higher-order/with-featured-image-notice.js
