+++ assets/src/components/higher-order/with-story-featured-image-notice.js
