+++ assets/src/components/index.js
@@ -22,8 +22,8 @@ export { default as withAttributes } from './higher-order/with-attributes';
 export { default as withBlockName } from './higher-order/with-block-name';
 export { default as withHasSelectedInnerBlock } from './higher-order/with-has-selected-inner-block';
 export { default as withPageNumber } from './higher-order/with-page-number';
-export { default as withFeaturedImageNotice } from './with-featured-image-notice';
-export { default as withStoryFeaturedImageNotice } from './with-story-featured-image-notice';
+export { default as withFeaturedImageNotice } from './higher-order/with-featured-image-notice';
+export { default as withStoryFeaturedImageNotice } from './higher-order/with-story-featured-image-notice';
 export { default as withEditFeaturedImage } from './with-edit-featured-image';
 export { default as withIsReordering } from './higher-order/with-is-reordering';
 export { default as withSelectedBlock } from './higher-order/with-selected-block';
